requirements.txt
fastapi
uvicorn
pandas
matplotlib
python-dotenv
requests
tabulate
mistralai>=1.0.0

langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-experimental==0.0.62
langsmith==0.1.147